Move performance tests for String.Contains to netcoreapp
Fix #20846
