+++ src/System.Runtime/tests/Performance/Perf.String.cs
@@ -10,7 +10,7 @@ using Microsoft.Xunit.Performance;
 
 namespace System.Tests
 {
-    public class Perf_String
+    public partial class Perf_String
     {
         public static IEnumerable<object[]> TestStringSizes()
         {
@@ -88,19 +88,6 @@ namespace System.Tests
                 using (iteration.StartMeasurement())
                     for (int i = 0; i < 10000; i++)
                         testString.Contains(subString);
-        }
-
-        [Benchmark]
-        [MemberData(nameof(ContainsStringComparisonArgs))]
-        public void Contains(StringComparison comparisonType, int size)
-        {
-            PerfUtils utils = new PerfUtils();
-            string testString = utils.CreateString(size);
-            string subString = testString.Substring(testString.Length / 2, testString.Length / 4);
-            foreach (var iteration in Benchmark.Iterations)
-                using (iteration.StartMeasurement())
-                    for (int i = 0; i < 10000; i++)
-                        testString.Contains(subString, comparisonType);
         }
 
         [Benchmark]
@@ -463,7 +450,6 @@ namespace System.Tests
         public static IEnumerable<object[]> SplitArgs => Permutations(s_trimStrings, s_trimCharArrays, s_splitOptions);
         public static IEnumerable<object[]> CompareArgs => Permutations(s_comparePairs, s_compareOptions);
         public static IEnumerable<object[]> IndexOfArgs => Permutations(s_compareOptions);
-        public static IEnumerable<object[]> ContainsStringComparisonArgs => Permutations(s_compareOptions, TestStringSizes());
 
         [Benchmark]
         [MemberData(nameof(CaseArgs))]
